add cap on village resources

+++ backend/internal/models/config_model.go
@@ -26,7 +26,7 @@ type ResourceConfig struct {
 	Level              int    `json:"level" db:"level"`
 	ResourceType       string `json:"resource_type" db:"resource_type"`
 	MaxCapacity        int    `json:"max_capacity" db:"max_capacity"`
-	ResourcePerSec     int    `json:"resource_per_sec" db:"resource_per_sec"`
+	ResourcePerSec     float64 `json:"resource_per_sec" db:"resource_per_sec"`
 	UpgradeCost        int    `json:"upgrade_cost" db:"upgrade_cost"`
 	UpgradeCostType    string `json:"upgrade_cost_type" db:"upgrade_cost_type"`
 	UpgradeDurationSec int    `json:"upgrade_duration_sec" db:"upgrade_duration_sec"`
